backend/requirements.txt
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
Django==2.2.16
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
python-dotenv==0.21.0
django-environ==0.9.0
psycopg2==2.9.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter==2.4.0
djoser==2.1.0
